aded jest to package.json and refactored template folder
