+++ src/app/(web-template)/arthur/annonces/[id]/(detailAnnonce)/layout.tsx
