+++ src/app/(web-template)/arthur/newsletter/confirm/loading.tsx
