+++ src/app/(web-template)/arthur/newsletter/confirm/page.tsx
