+++ src/app/(web-template)/arthur/newsletter/delete/loading.tsx
