+++ src/app/(web-template)/arthur/newsletter/delete/page.tsx
