+++ src/app/(web-template)/theo/_components/TheoHero.tsx
