+++ src/app/(web-template)/theo/_components/TheoPropertiesSection.tsx
