+++ src/app/(web-template)/theo/_components/TheoPropertyCard.tsx
