+++ src/app/(web-template)/theo/_components/TheoPropertyGridCard.tsx
