+++ src/app/(web-template)/theo/_components/TheoSearchBar.tsx
