+++ src/app/(web-template)/theo/annonces/[id]/(detailAnnonce)/loading.tsx
