+++ src/app/(web-template)/theo/blog/[slug]/page.tsx
