+++ src/app/(web-template)/theo/blog/layout.tsx
