+++ src/app/(web-template)/theo/newsletter/confirm/loading.tsx
