+++ src/app/(web-template)/theo/newsletter/confirm/page.tsx
