+++ src/app/(web-template)/theo/newsletter/delete/loading.tsx
