+++ src/app/(web-template)/theo/newsletter/delete/page.tsx
